+++ CHANGELOG.md
@@ -7,6 +7,8 @@ the release.
 
 ## Unreleased
 
+* [shipping] Add host resource detection to enrich SDK resource metadata
+  ([#3581](https://github.com/open-telemetry/opentelemetry-demo/pull/3581))
 * [frontend-proxy] Use the asynchronous c-ares DNS resolver for Envoy upstream
   clusters instead of the blocking `getaddrinfo` resolver. With `getaddrinfo`, a
   slow or unanswered DNS lookup for an upstream that is not running (e.g. the